Various fixes to SCP's algorithm
Most important change: created a replyMessage so an externalized node
can still answer someone. This ensures our proposer will reach
externalized. But it doesn't ensure everyone will, but that is fine.

Also a bug on the areBallotsEqual.

And various conditions on the steps, so node don't fall from
externalized to finished/prepare.